test: create root store for testing
